chore: rewrite the rev-list parsing to be safe and extensible
breaking changes:
* the date format is now iso8601 per github events
* author.name and author.email are now null if they're not inferrable
  from the commit.  For reintegrating the name/email on the consuming side,
  null is simpler than trying to check if == 'unknown'.

The previous logic relied on \n\n splitting, but people *can*
put that in there.  Instead this pathway now just gets the raw fields
straight out of git without formatting so the parser cannot choke
on oddball commit messages.

Additionally shove in committer data, parents (so merge
is known), and trailers as a list of dicts.

Signed-off-by: Brian Harring <ferringb@gmail.com>